+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/query/IDocumentQueryClient.java
@@ -2,15 +2,16 @@
 // Licensed under the MIT License.
 package com.azure.cosmos.implementation.query;
 
+import com.azure.cosmos.ConsistencyLevel;
 import com.azure.cosmos.implementation.DocumentClientRetryPolicy;
+import com.azure.cosmos.implementation.IRetryPolicyFactory;
 import com.azure.cosmos.implementation.OperationType;
 import com.azure.cosmos.implementation.ResourceType;
-import com.azure.cosmos.implementation.caches.IPartitionKeyRangeCache;
-import com.azure.cosmos.implementation.caches.RxCollectionCache;
-import com.azure.cosmos.ConsistencyLevel;
-import com.azure.cosmos.implementation.IRetryPolicyFactory;
 import com.azure.cosmos.implementation.RxDocumentServiceRequest;
 import com.azure.cosmos.implementation.RxDocumentServiceResponse;
+import com.azure.cosmos.implementation.caches.IPartitionKeyRangeCache;
+import com.azure.cosmos.implementation.caches.RxCollectionCache;
+import com.azure.cosmos.models.FeedResponse;
 import reactor.core.publisher.Mono;
 
 import java.util.function.BiFunction;
@@ -41,13 +42,13 @@ public interface IDocumentQueryClient {
 
     /**
      * TODO: this should be async returning observable
-     * @return 
+     * @return
      */
     ConsistencyLevel getDefaultConsistencyLevelAsync();
 
     /**
      * TODO: this should be async returning observable
-     * @return 
+     * @return
      */
     ConsistencyLevel getDesiredConsistencyLevelAsync();
 
@@ -61,6 +62,10 @@ public interface IDocumentQueryClient {
         final Supplier<DocumentClientRetryPolicy> retryPolicyFactory,
         final RxDocumentServiceRequest req,
         final BiFunction<Supplier<DocumentClientRetryPolicy>, RxDocumentServiceRequest, Mono<T>> feedOperation);
+
+    <T> Mono<FeedResponse<T>> getFeedResponseMonoWithTimeout(
+        final RxDocumentServiceRequest req,
+        final Mono<FeedResponse<T>> responseMono);
 
     /// <summary>
     /// A client query compatibility mode when making query request.