+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/CosmosQueryRequestOptions.java
@@ -4,7 +4,6 @@
 package com.azure.cosmos.models;
 
 import com.azure.cosmos.ConsistencyLevel;
-import com.azure.cosmos.CosmosDiagnostics;
 import com.azure.cosmos.CosmosDiagnosticsThresholds;
 import com.azure.cosmos.CosmosEndToEndOperationLatencyPolicyConfig;
 import com.azure.cosmos.implementation.Configs;
@@ -16,7 +15,6 @@ import com.azure.cosmos.implementation.spark.OperationContextAndListenerTuple;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import java.time.Duration;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -59,8 +57,6 @@ public class CosmosQueryRequestOptions {
     private String queryName;
     private CosmosEndToEndOperationLatencyPolicyConfig cosmosEndToEndOperationLatencyPolicyConfig;
     private List<String> excludeRegions;
-    private List<CosmosDiagnostics> cancelledRequestDiagnosticsTracker = new ArrayList<>();
-
     /**
      * Instantiates a new query request options.
      */
@@ -104,7 +100,6 @@ public class CosmosQueryRequestOptions {
         this.thresholds = options.thresholds;
         this.cosmosEndToEndOperationLatencyPolicyConfig = options.cosmosEndToEndOperationLatencyPolicyConfig;
         this.excludeRegions = options.excludeRegions;
-        this.cancelledRequestDiagnosticsTracker = options.cancelledRequestDiagnosticsTracker;
     }
 
     void setOperationContextAndListenerTuple(OperationContextAndListenerTuple operationContextAndListenerTuple) {
@@ -695,14 +690,6 @@ public class CosmosQueryRequestOptions {
         return this;
     }
 
-    List<CosmosDiagnostics> getCancelledRequestDiagnosticsTracker() {
-        return this.cancelledRequestDiagnosticsTracker;
-    }
-
-    void setCancelledRequestDiagnosticsTracker(List<CosmosDiagnostics> cancelledRequestDiagnosticsTracker) {
-        this.cancelledRequestDiagnosticsTracker = cancelledRequestDiagnosticsTracker;
-    }
-
     ///////////////////////////////////////////////////////////////////////////////////////////
     // the following helper/accessor only helps to access this class outside of this package.//
     ///////////////////////////////////////////////////////////////////////////////////////////
@@ -829,18 +816,6 @@ public class CosmosQueryRequestOptions {
                 @Override
                 public CosmosEndToEndOperationLatencyPolicyConfig getEndToEndOperationLatencyPolicyConfig(CosmosQueryRequestOptions options) {
                     return options.getEndToEndOperationLatencyConfig();
-                }
-
-                @Override
-                public List<CosmosDiagnostics> getCancelledRequestDiagnosticsTracker(CosmosQueryRequestOptions options) {
-                    return options.getCancelledRequestDiagnosticsTracker();
-                }
-
-                public void setCancelledRequestDiagnosticsTracker(
-                    CosmosQueryRequestOptions options,
-                    List<CosmosDiagnostics> cancelledRequestDiagnosticsTracker) {
-
-                    options.setCancelledRequestDiagnosticsTracker(cancelledRequestDiagnosticsTracker);
                 }
 
                 @Override